+++ CHANGELOG.md
@@ -148,6 +148,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fixed ghost holograms caused by reloading BQ
 - fixed deadlock(Server crash) in Conversations with a large amount of npc and player options with a large amount of conditions
 - fixed door event not working correctly
+- fixed `1 give` command exceptions
 ### Security
 - fixed issue, where objectives that count things are out of sync with the database. This has also affected BungeeCord support